Handle unicode output from nova properly
If nova returns unicode characters to the terminal they aren't displayed correctly and a
unicode error is thrown.  This fix should connect the stdout/stderr of the subprocess to
that of the parent and allow for unicode characters to be displayed properly.

Thanks to jkoelker for the help on this one.